Reorganise imports to match other examples
